Default pen settings, fix crash, add documentation
Now, the addStroke API call will default to current color and size
settings if they are missing from the API call.

Fixed potential crash with refreshPage function by adding a check for a
selected page. Throws a warning if it can't find one.

Add more documentation for addStroke and addSpline